backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
google-cloud-storage>=2.13.0
python-dotenv>=1.0.0
tenacity>=8.2.0
firebase-admin>=6.2.0
